+++ natservice/api/src/main/java/org/opendaylight/netvirt/natservice/api/SnatServiceManager.java
@@ -7,12 +7,12 @@
  */
 package org.opendaylight.netvirt.natservice.api;
 
-import java.math.BigInteger;
 import java.util.concurrent.ExecutionException;
 import org.eclipse.jdt.annotation.Nullable;
 import org.opendaylight.genius.infra.Datastore.Configuration;
 import org.opendaylight.genius.infra.TypedReadWriteTransaction;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.natservice.rev160111.ext.routers.Routers;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 public interface SnatServiceManager {
 
@@ -29,7 +29,7 @@ public interface SnatServiceManager {
     }
 
     void notify(TypedReadWriteTransaction<Configuration> confTx, Routers router, @Nullable Routers oldRouter,
-        BigInteger primarySwitchId, @Nullable BigInteger dpnId, Action action)
+                Uint64 primarySwitchId, @Nullable Uint64 dpnId, Action action)
         throws ExecutionException, InterruptedException;
 
 }
